Updates for Grandstream compatibility
Modify codec negotiation to work with Grandstream
Move example code out of call_phone.py
